% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exonic_lengths.R
\name{get_exonic_lengths}
\alias{get_exonic_lengths}
\title{Returns the cumulated exonic length for each gene}
\usage{
get_exonic_lengths(org = "hg19")
}
\arguments{
\item{org}{organism}
}
\description{
Returns the cumulated exonic length for each gene
}
