% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivities.R
\name{get_lambda_max}
\alias{get_lambda_max}
\title{Compute the max glmnet lambda penalization across several samples}
\usage{
get_lambda_max(Y, X, alpha)
}
\arguments{
\item{Y}{observation/response vector}

\item{X}{predictor matrix}

\item{alpha}{mixing parameter for the penalty term (0 -> ridge, 1 -> lasso)}
}
\value{
the smallest value of lambda for which all coefficients are set to 0 across all samples
}
\description{
Computes the max value of lambda such that all coefficients
remain 0 when running a penalized linear regression of Y onto X with the
glmnet package. Based on the coordinate descent udpate formula.
}
