% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tpm.R
\name{get_tpm}
\alias{get_tpm}
\title{computes TPMs from a table of counts. Use get_tpm_local when possible to save time}
\usage{
get_tpm(x, org = c("hg19", "mm9", "rheMac8"))
}
\arguments{
\item{x}{table with raw counts, ensembl id as rownames}

\item{org}{organism, one of c('hg19', 'mm9', 'rheMac8')}
}
\value{
a TPM matrix with ensembl ID as rownames, samples as columns
}
\description{
The count table should contain genes as rows and samples as columns
}
