% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivities.R
\name{getActivities}
\alias{getActivities}
\title{Estimate activities (transcription factors, TE subfamilies) using the elastic net regularization}
\usage{
getActivities(
  expr,
  X,
  alpha = 0.1,
  k = 5,
  nlambda = 100,
  ncores = 1,
  log_id = ""
)
}
\arguments{
\item{expr}{row-centered TPM values in a promoter/gene (rows) x sample (columns) matrix.}

\item{X}{per promoter/gene (rows) motif (columns) counts. Column-centered. Also referred to as N.}

\item{alpha}{trades off between ridge (alpha = 0) and lasso (alpha = 1) penalties}

\item{k}{number of folds for cross-validation}

\item{nlambda}{number of lambda values in the grid from lambda max to lambda min}

\item{ncores}{number of cores for multiprocessing (1 -> no multiprocessing)}
}
\value{
a list object with the matrix of estimated activities, and diagnostics (e.g r2 etc)
}
\description{
Estimates activities by regressing each
column of the expression matrix (1 col per sample) onto the per promoter
motif site counts. Multiple linear regression with the
elastic net regularization is performed by the glmnet package.
We aim to find the value of lambda which minimizes the error
over all samples. Therefore, we must provide the same grid of lambda values
to each regression problem (one per sample). The max value of lambda is
defined as the maximum value of lambda_max across samples.
lambda_max for a single sample is given by the formula:
lambda_max = 1/(nobs*alpha)*max_s(abs(dot product(Z(x_s), y_s))) where Z
is a function that centers AND standardizes its input.
}
