% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignifActivities.R
\name{getSignifActivities}
\alias{getSignifActivities}
\title{Estimates and returns the activities of TE subfamilies by multiple linear regression}
\usage{
getSignifActivities(E_centered, N, treatment_group, control_group)
}
\arguments{
\item{E_centered}{matrix of gene expression values with genes as rows and samples as columns}

\item{N}{predictor matrix with genes as rows and TE subfamilies as columns}

\item{treatment_group}{vector with the name of treated samples (correspond to columns in E_centered)}

\item{control_group}{same for control samples}
}
\value{
list object with estimated activities and useful diagnostics (e.g R2)
}
\description{
Each activity coefficient is tested agains the null hypothesis that it is equal to zero
A Benjamini-Hochberg procedure to control the FDR is employed.
}
