% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_from_tsv.R
\name{N_from_tsv}
\alias{N_from_tsv}
\title{fast loading of N from a .tsv, using data.table}
\usage{
N_from_tsv(filename)
}
\arguments{
\item{filename}{path to tsv file specifying N, with genes as rows and TE subfamilies as columns}
}
\value{
matrix N
}
\description{
loads N specified in a .tsv file, and returns it as a matrix ready for use in preprocess_E_N_for_activities()
}
