% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signif_subfams.R
\name{plot_signif_subfams}
\alias{plot_signif_subfams}
\title{Volcano plot with TE subfamily activities on the x axis, activity coefficient significance on the y axis}
\usage{
plot_signif_subfams(
  res,
  alpha,
  n_label = NULL,
  main = NULL,
  n_subplots = 1,
  idx = 1,
  plot_signif_thresh = FALSE,
  saturate_padj_at = 1e-150
)
}
\arguments{
\item{res}{result file returned by getSignifActivities.R}

\item{alpha}{BH-adjusted threshold for significance (black circles)}

\item{n_label}{number of subfamilies to label on the plot, starting from the most significant subfamily}

\item{main}{title of the plot}

\item{n_subplots}{when plotting several plots on the same figure (e.g par(mar=c(2, 2))), we only reset par() after the last subplot is drawn}

\item{idx}{belongs to the interval \link{1, n_subplots} and triggers the reset of par() when it equals n_subplots}

\item{plot_signif_thresh}{whether to explicitely plot the significance threshold as a dotted line}

\item{saturate_padj_at}{representative p-value beyond which any smaller adjusted p-value is "saturated" to}
}
\description{
Black dots are significant activities, grey dots are non significant activities.
The 95\% confidence interval are plotted for all significant activities as grey bars.
}
