% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signif_subfams_functional.R
\name{plot_signif_subfams_functional}
\alias{plot_signif_subfams_functional}
\title{Volcano plot with TE subfamily activities on the x axis, activity coefficient significance on the y axis, with functional vs non functional activities for the selected subfamilies}
\usage{
plot_signif_subfams_functional(
  res,
  res_functional,
  subfams_to_plot,
  main = NULL,
  n_subplots = 1,
  idx = 1,
  subfams_to_label = NULL,
  suffix = "_functional"
)
}
\arguments{
\item{res}{result file returned by getSignifActivities.R}

\item{res_functional}{result file returned by \code{getSignifActivities()} using the functional version of N}

\item{subfams_to_plot}{subfamilies for which to plot the functional vs non functional activities}

\item{main}{title of the plot}

\item{n_subplots}{when plotting several plots on the same figure (e.g par(mar=c(2, 2))), we only reset par() after the last subplot is drawn}

\item{idx}{belongs to the interval \link{1, n_subplots} and triggers the reset of par() when it equals n_subplots}

\item{subfams_to_label}{subfamilies for which a label is written on the plot, next to the functional activity point (red). Defaults to \code{subfams_to_plot}}

\item{suffix}{character string denoting the functional fractions of subfamilies}
}
\description{
Black dot: activity of the entire subfamily. Red dot: activity of the functional fraction of the subfamily. Blue dot: activity of the non-functional fraction of the subfamily.
}
