% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivities.R
\name{plotActivities}
\alias{plotActivities}
\title{Scatter plot of TF activities per sample}
\usage{
plotActivities(
  act,
  TF,
  metadata,
  sort_by,
  decreasing = F,
  palette_type = "sequential",
  palette_add_black = T,
  palette_black_idx = 1,
  title = NULL,
  act_lim = NULL
)
}
\arguments{
\item{act}{TF x samples matrix, with row and col names set}

\item{TF}{string for TF of interest, e.g "MYC"}

\item{metadata}{samples x metadata dataframe. Metadata should be set as factors}

\item{sort_by}{string corresponding to a column of metadata}

\item{decreasing}{bool to sort samples by metadata\link{sort_by} in decr. order.}

\item{palette_type}{string, one of c("sequential", "diverging", "categorical")}

\item{palette_add_black}{bool to enable setting a color to black (useful for baselines)}

\item{palette_black_idx}{category to plot in black}

\item{title}{title if it is to be different than the TF name}

\item{act_lim}{limits of activities to plot, useful to get comparable y axes}
}
\description{
Scatter plot of TF activities per sample
}
