% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_E_N_for_activities.R
\name{preprocess_E_N_for_activities}
\alias{preprocess_E_N_for_activities}
\title{Preprocess an RNA seq count table for TE subfamily activity estimation}
\usage{
preprocess_E_N_for_activities(
  count_table,
  N,
  count_threshold = c(10, 1),
  N_threshold = 150,
  pseudocount_quantile = 0.05,
  log_tpm_plot_path,
  preds_protected = NULL,
  redownload_exons = FALSE,
  exons_dir = "../data/temp/"
)
}
\arguments{
\item{count_table}{gene count matrix (rows are genes, columns are samples).}

\item{N}{per gene occurences of TE integrants. Genes are rows, TE subfamilies are columns}

\item{N_threshold}{minimum sum over all genes that a TE subfamily has to reach in order to be kept as a predictor in matrix N}

\item{pseudocount_quantile}{quantile of non-zero values used to compute per-sample pseudocounts prior to log-transformation}

\item{log_tpm_plot_path}{path where histograms of expression values (in logged TPM) are drawn (useful to check quasi-normality of expression values)}

\item{preds_protected}{vector containing the names of predictors that cannot be filtered out based on N_threshold}

\item{redownload_exons}{TRUE: exons are redownloaded to compute TPMs. FALSE: exons are not redownloaded to compute TPMs.}

\item{exons_dir}{directory where gene models are saved to avoid a redownload}

\item{count_threhold}{c(a,b) with a the minimum number of counts in at least b genes to keep a gene present in E}

\item{import_path}{path pointing to the other R functions required for the pre-processing}
}
\value{
a list object with a centered expression matrix and a filtered N matrix
}
\description{
Filters, normalizes and processes a table of raw counts with genes as rows and samples as columns to
the expression matrix E. Builds a corresponding N matrix with remaining genes as rows and TEs as columns.
}
