% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_from_fstat.R
\name{pval_from_fstat}
\alias{pval_from_fstat}
\title{Returns the pvalue corresponding to the f-statistic of the complete versus the null model (intercept only)}
\usage{
pval_from_fstat(res)
}
\arguments{
\item{res}{result returned by getSignifActivities}
}
\value{
p-vaue for overall significance test of the linear model
}
\description{
Returns the pvalue corresponding to the f-statistic of the complete versus the null model (intercept only)
}
