% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_center.R
\name{row_center}
\alias{row_center}
\title{Centers the matrix X by rows}
\usage{
row_center(X)
}
\arguments{
\item{X}{matrix with rows to be centered.}
}
\value{
row-centered matrix X
}
\description{
Centers the matrix X by rows
}
