% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_N_to_functional.R
\name{split_N_to_functional}
\alias{split_N_to_functional}
\title{split a cis-regulatory susceptibility matrix N into functional and non functional fractions of TE subfamilies}
\usage{
split_N_to_functional(
  N,
  N_functional,
  threshold = 100,
  subfams_protected = c()
)
}
\arguments{
\item{N}{pre-filtered input N matrix}

\item{N_functional}{unfiltered functional matrix corresponding to N (e.g same bandwidth parameter L)}

\item{threshold}{minimum sum of cis-regulatory weights over genes at both the non-funtional and funtional fractions for a subfamily to be split}

\item{subfams_protected}{will be split irrespective of \code{threshold}}
}
\value{
functional equivalent of the input matrix \code{N}
}
\description{
starting from a specific N matrix that was already filtered for specific genes (e.g defined through \code{preprocess_E_N_for_activities}), split subfamilies into functional vs non_functional fractions according to a source functional N matrix defined for all genes. Subfamilies are only split if both their functional and non functional sum of cis-regulatory weights are above the threshold.
}
